Added ratelimiting to 401 Unathorized responses to prevent abuse of the server for use in DDoS attacks via traffic reflection and amplification.
This patch works by counting the amount of requests that result in a 401
Unathorized response and limiting them by IP Address if the occured in
a specified window of time.

ur_addr_map* functions were extended and wrapped to enable ioa_addr objects
with and without port numbers. In our usecase we need to be able use the
*no_port variants when working with `ioa_addr` types.

Added new command-line options:
  --no-ratelimit-401 - Disables ratelimiting of 401 Responses
  --ratelimit-401-requests-per-window - Sets the amount of requests that result
                                      a 401 response per ratelimt window
  --ratelimit-401-window-seconds - Sets the size in seconds of the ratelimt
                                 window.